pyproject.toml
[project]
name = "pdf-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.7",
    "fastapi>=0.124.4",
    "ipykernel>=7.1.0",
    "langchain>=1.1.3",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
